LayoutTests/svg/animations/animations-paused-disconnected-iframe-expected.txt

Tests that SVG animations are properly paused in disconnected iframes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS frameWindow.internals.areSVGAnimationsPaused is false
PASS grandChildWindow.internals.areSVGAnimationsPaused is false
Removing iframe from the document.
PASS frameWindow.internals.areSVGAnimationsPaused is true
PASS grandChildWindow.internals.areSVGAnimationsPaused is true
Add iframe back into the document.
PASS frame.contentWindow.internals.areSVGAnimationsPaused is false
PASS frame.contentDocument.getElementById('grandChildFrame').contentWindow.internals.areSVGAnimationsPaused is false
PASS successfullyParsed is true

TEST COMPLETE
 
